Move concurrent to core. Add tricky highlight test
